// src/webparts/detailsCustomList/interfaces/ISelectedItem.ts
export interface ISelectedItem {
  selectedItemId: string | number;
  selectedItemName: string;
  serverRelativeUrl: string;
  selectedItemUrlDownload: string;
  selectedItemUrlOpenInBrowser: string;
  selectedItemDocId: string;
  selectedItemExt: string;
  selectedItemUniqueId: string;
  selectedList: string;
}